Adjusted semantics of *Complete.GetFramebuffer()
These seem to be const references in the C++ implementation. Likely, these will
refer to wrapped internal Framebuffer instances that were explicitly created and
passed to the Complete instance. As such, no finalizer is attached to these
references.

If this analysis is correct, it is possible to retain dangling pointer references
to these Framebuffer instances if the parent reference is finalized. To correct
this otherwise would require some kind of reference counting.