Add SWIG wrapping of Terminal::DrawState and add it to pkg/mosh/terminal
Added a Go type pkg/mosh/terminal/DrawState with GetWidth() and
GetHeight() methods. Also added shorthand methods to the pkg
version of Terminal::Framebuffer with the same names.

Need to be able to read the current window dimensions of a framebuffer
to detect size changes in response to processing Resize events.